Better handling of null/undefined as root of responses with Node.
Node passes messges using JSON strings (it seems), so setting a type to "undefined"
seems to make its property to completely disppear from the message. We add another
couple of properties there ("undefined" and also "null" just to be on the safe side)
that are then re-parsed on the server side.